[Refactor] process test don't depend on kafka
