+++ engine/flink/kafka-util/src/test/resources/META-INF/services/pl.touk.nussknacker.engine.flink.api.typeinformation.TypingResultAwareTypeInformationCustomisation
@@ -1,0 +1,2 @@
+pl.touk.nussknacker.engine.kafka.source.InputMetaTypeInformationCustomisation
+pl.touk.nussknacker.engine.kafka.source.SampleKeyTypeInformationCustomisation
