--- engine/flink/process/src/test/resources/META-INF/services/pl.touk.nussknacker.engine.flink.api.typeinformation.TypingResultAwareTypeInformationCustomisation
@@ -1,2 +1,0 @@
-pl.touk.nussknacker.engine.kafka.source.InputMetaTypeInformationCustomisation
-pl.touk.nussknacker.engine.process.source.SampleKeyTypeInformationCustomisation
