+++ engine/flink/process/src/test/scala/pl/touk/nussknacker/engine/process/helpers/ProcessTestHelpers.scala
@@ -1,6 +1,5 @@
 package pl.touk.nussknacker.engine.process.helpers
 
-import com.typesafe.config.ConfigValueFactory.fromAnyRef
 import com.typesafe.config.{Config, ConfigFactory}
 import org.apache.flink.api.common.ExecutionConfig
 import org.apache.flink.streaming.api.scala._
@@ -11,11 +10,11 @@ import pl.touk.nussknacker.engine.api.dict.DictInstance
 import pl.touk.nussknacker.engine.api.dict.embedded.EmbeddedDictDefinition
 import pl.touk.nussknacker.engine.api.exception.ExceptionHandlerFactory
 import pl.touk.nussknacker.engine.api.process.{ProcessObjectDependencies, _}
+import pl.touk.nussknacker.engine.api.signal.ProcessSignalSender
 import pl.touk.nussknacker.engine.api.typed.TypedMap
 import pl.touk.nussknacker.engine.flink.api.process._
 import pl.touk.nussknacker.engine.flink.test.FlinkSpec
 import pl.touk.nussknacker.engine.graph.EspProcess
-import pl.touk.nussknacker.engine.kafka.KafkaConfig
 import pl.touk.nussknacker.engine.process.compiler.FlinkProcessCompiler
 import pl.touk.nussknacker.engine.process.helpers.SampleNodes._
 import pl.touk.nussknacker.engine.process.registrar.FlinkProcessRegistrar
@@ -31,7 +30,6 @@ trait ProcessTestHelpers extends FlinkSpec { self: Suite =>
                              processVersion: ProcessVersion = ProcessVersion.empty,
                              parallelism: Int = 1): Unit = {
       val config = ConfigFactory.load()
-        .withValue("kafka.kafkaAddress", fromAnyRef("http://notexist.pl"))
       val creator: ProcessConfigCreator = ProcessTestHelpers.prepareCreator(data, config)
 
       val env = flinkMiniCluster.createExecutionEnvironment()
@@ -58,21 +56,11 @@ trait ProcessTestHelpers extends FlinkSpec { self: Suite =>
       MockService.clear()
       env.withJobRunning(process.id)(actionToInvokeWithJobRunning)
     }
-
-    def invokeWithKafka(process: EspProcess,
-                        config: Config,
-                        processVersion: ProcessVersion = ProcessVersion.empty,
-                        parallelism: Int = 1)(actionToInvokeWithJobRunning: => Unit): Unit = {
-      val creator: ProcessConfigCreator = ProcessTestHelpers.prepareCreator(Nil, config)
-      invoke(process, creator, config, processVersion, parallelism, actionToInvokeWithJobRunning)
-    }
   }
 
 }
 
 object ProcessTestHelpers {
-
-  val signalTopic = "signals1"
 
   def prepareCreator(data: List[SimpleRecord], config: Config): ProcessConfigCreator = new ProcessConfigCreator {
 
@@ -87,7 +75,6 @@ object ProcessTestHelpers {
     override def sourceFactories(processObjectDependencies: ProcessObjectDependencies): Map[String, WithCategories[FlinkSourceFactory[_]]] = Map(
       "input" -> WithCategories(SampleNodes.simpleRecordSource(data)),
       "intInputWithParam" -> WithCategories(new IntParamSourceFactory(new ExecutionConfig)),
-      "kafka-keyvalue" -> WithCategories(new KeyValueKafkaSourceFactory(processObjectDependencies)),
       "genericParametersSource" -> WithCategories(GenericParametersSource),
       "genericSourceWithCustomVariables" -> WithCategories(GenericSourceWithCustomVariables)
     )
@@ -108,8 +95,6 @@ object ProcessTestHelpers {
       "customContextClear" -> WithCategories(CustomContextClear),
       "sampleJoin" -> WithCategories(CustomJoin),
       "joinBranchExpression" -> WithCategories(CustomJoinUsingBranchExpressions),
-      "signalReader" -> WithCategories(CustomSignalReader),
-      "transformWithTime" -> WithCategories(TransformerWithTime),
       "transformWithNullable" -> WithCategories(TransformerWithNullableParam),
       "optionalEndingCustom" -> WithCategories(OptionalEndingCustom),
       "genericParametersNode" -> WithCategories(GenericParametersNode),
@@ -132,11 +117,7 @@ object ProcessTestHelpers {
       ExpressionConfig(globalProcessVariables, List.empty, dictionaries = Map(dictId -> WithCategories(dictDef)))
     }
 
-    override def signals(processObjectDependencies: ProcessObjectDependencies): Map[String, WithCategories[TestProcessSignalFactory]] = {
-      val kafkaConfig = KafkaConfig.parseConfig(processObjectDependencies.config)
-      Map("sig1" ->
-        WithCategories(new TestProcessSignalFactory(kafkaConfig, signalTopic)))
-    }
+    override def signals(processObjectDependencies: ProcessObjectDependencies): Map[String, WithCategories[ProcessSignalSender]] = Map.empty
 
     override def buildInfo(): Map[String, String] = Map.empty
   }
