+++ engine/flink/test-util/src/main/java/pl/touk/nussknacker/engine/process/SimpleJavaEnum.java
